patch-video: Direct3D11 texture read/write/copy: add tests + improve comments

+++ video/src/d3d11/texture_reader_writer.hpp
@@ -47,7 +47,7 @@ Implementation included in renderer.cpp
     else { // no padding -> straight copy
       if (sliceBytes > mapped.DepthPitch)
         sliceBytes = mapped.DepthPitch; // protect against user calculation errors -> avoids crashes
-      memcpy(outputData, mapped.pData, static_cast<size_t>(sliceBytes*depthSliceCount));
+      memcpy(outputData, mapped.pData, (size_t)sliceBytes*(size_t)depthSliceCount);
     }
   }
